Revert "TASK: Use t3 env and separate sanity queries from client"
