--- .env.example
@@ -1,4 +1,0 @@
-NEXT_PUBLIC_SANITY_PROJECT_ID=
-NEXT_PUBLIC_SANITY_DATASET=
-NEXT_PUBLIC_SANITY_API_VERSION=
-SANITY_REVALIDATE_SECRET=
