+++ package.json
@@ -12,7 +12,6 @@
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-slot": "^1.0.2",
-    "@t3-oss/env-nextjs": "^0.8.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "geist": "^1.2.1",
@@ -24,8 +23,7 @@
     "server-only": "^0.0.1",
     "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
-    "vaul": "^0.8.9",
-    "zod": "^3.22.4"
+    "vaul": "^0.8.9"
   },
   "devDependencies": {
     "@types/node": "^20.11.6",
