+++ src/app/(personal)/timeline/page.tsx
@@ -3,7 +3,7 @@ import { PageContent } from '@/components/page-content';
 import { PageWrapper } from '@/components/page-wrapper';
 import { ScrollView } from '@/components/scroll-view';
 import { PageBuilder } from '@/components/page-builder';
-import { sanity } from '@/lib/sanity.client';
+import { sanity } from '@/lib/sanity';
 import { Header } from '@/components/header';
 
 export default async function Timeline() {
