+++ src/components/timeline.tsx
@@ -35,7 +35,7 @@ export function Timeline({ content }: Props) {
                   </div>
                   <TimelineItem
                     timelineItem={timelineItem}
-                    isPriorityImage={index < 1}
+                    isPriorityImage={index < 3}
                   />
                 </div>
               ))}
